Add basic calculator
Just perform basic function according to operator variable which represents as 'char', if the value of char change, when input first and second number it will change operator of the variable.
i.e. 

i j k = x

WHERE i is the first variable
WHERE j is the operator
WHERE x is the result of three variables performing operation together according to the j variable as operator.